pkg/operator: move internal tests to blackbox
Closes #141
